refactor(components/molecule/select): add import file extensions
