+++ components/molecule/select/demo/components/MoleculeSelectUseEffect.js
@@ -3,12 +3,10 @@ import {useState, useEffect} from 'react'
 import PropTypes from 'prop-types'
 
 import {withStateValue} from '@s-ui/hoc'
-
-import MoleculeSelect from 'components/molecule/select/src'
-
 import MoleculeSelectOption from '@s-ui/react-molecule-dropdown-option'
 
-import {IconArrowDown} from '../Icons/'
+import MoleculeSelect from '../../src/index.js'
+import {IconArrowDown} from '../Icons/index.js'
 import regions from '../data/regions.json'
 
 const MoleculeSelectWithState = withStateValue(MoleculeSelect)
