+++ components/molecule/select/src/components/MultipleSelection.js
@@ -1,6 +1,7 @@
 import MoleculeDropdownList from '@s-ui/react-molecule-dropdown-list'
 import MoleculeInputTags from '@s-ui/react-molecule-input-tags'
-import MoleculeInputSelect from './MoleculeInputSelect'
+
+import MoleculeInputSelect from './MoleculeInputSelect.js'
 
 const MoleculeSelectFieldMultiSelection = props => {
   /* eslint-disable react/prop-types */
