+++ components/molecule/select/src/components/config.js
@@ -1,0 +1,7 @@
+import {BASE_CLASS as ROOT_BASE_CLASS} from '../config.js'
+
+export const BASE_CLASS = `${ROOT_BASE_CLASS}-inputSelect`
+export const CLASS_CONTAINER = `${BASE_CLASS}-container`
+export const CLASS_ARROW = `${BASE_CLASS}-arrow`
+export const CLASS_ARROW_DOWN = `${CLASS_ARROW}--down`
+export const CLASS_ARROW_UP = `${CLASS_ARROW}--up`
